module biometric refactoring
