--- src/modules/biometric-auth/camera.js
@@ -1,104 +1,0 @@
-/**
- * @license
- * Copyright 2022 Google LLC. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * =============================================================================
- */
-
-import {VIDEO_SIZE} from './shared/params';
-import {drawResults, isMobile} from './shared/util';
-
-export class Camera {
-  constructor() {
-    this.video = document.getElementById('video');
-    // this.canvas = document.getElementById('output');
-    // this.ctx = this.canvas.getContext('2d');
-    this.stream = null;
-  }
-
-  /**
-   * Initiate a Camera instance and wait for the camera stream to be ready.
-   * @param cameraParam From app `STATE.camera`.
-   */
-  static async setupCamera(cameraParam) {
-    if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia) {
-      throw new Error(
-          'Browser API navigator.mediaDevices.getUserMedia not available');
-    }
-
-    const {targetFPS, sizeOption} = cameraParam;
-    const $size = VIDEO_SIZE[sizeOption];
-    const videoConfig = {
-      'audio': false,
-      'video': {
-        facingMode: 'user',
-        // Only setting the video to a specified size for large screen, on
-        // mobile devices accept the default size.
-        // width: isMobile() ? VIDEO_SIZE['360 X 270'].width : $size.width,
-        // height: isMobile() ? VIDEO_SIZE['360 X 270'].height : $size.height,
-        // frameRate: {
-        //   ideal: targetFPS,
-        // },
-
-        // width: { min: 1280, ideal: 1920, max: 2560 },
-        // height: { min: 720, ideal: 1080, max: 1440 },
-
-        width: 600,
-        height: 400,
-
-        frameRate: { ideal: 30, max: 60 }
-
-      },
-    };
-
-    const stream = await navigator.mediaDevices.getUserMedia(videoConfig);
-
-    const camera = new Camera();
-    camera.video.srcObject = stream;
-    camera.stream = stream;
-
-    await new Promise((resolve) => {
-      camera.video.onloadedmetadata = () => {
-        resolve();
-      };
-    });
-
-    camera.video.play();
-
-    // const videoWidth = camera.video.videoWidth;
-    // const videoHeight = camera.video.videoHeight;
-    // Must set below two lines, otherwise video element doesn't show.
-    // camera.video.width = videoWidth;
-    // camera.video.height = videoHeight;
-
-    // camera.canvas.width = videoWidth;
-    // camera.canvas.height = videoHeight;
-    // const canvasContainer = document.querySelector('.canvas-wrapper');
-    // canvasContainer.style = `width: 640px; height: 480px`;
-
-    // Because the image from camera is mirrored, need to flip horizontally.
-    // camera.ctx.translate(640, 0);
-    // camera.ctx.scale(-1, 1);
-
-    return camera;
-  }
-
-  // drawCtx() {
-  //   this.ctx.drawImage(
-  //       this.video, 0, 0, 640, 480);
-  // }
-
-  // drawResults(faces, triangulateMesh, boundingBox) {
-  //   drawResults(this.ctx, faces, triangulateMesh, boundingBox);
-  // }
-}
