+++ src/modules/biometric-auth/helpers/calculateTasksProgress.js
